+++ src/index.ts
@@ -140,6 +140,7 @@ export * from "./components/Tooltip";
 export * from "./components/Select";
 export * from "./components/Combobox";
 export * from "./components/ThemedNumberInput";
+export * from "./components/CurrencyInput";
 export * from "./components/Toast";
 export * from "./components/ValueRenderer";
 export * from "./components/DiffPair";